+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [v0.5.0]
 ## Added
 - General transformers
 - Match-merge transformers
@@ -17,5 +19,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Breaking: changed interface names to match conventions
 
 
-[Unreleased]: https://github.com/smartive/proc-that/compare/v0.4.0...master
+[Unreleased]: https://github.com/smartive/proc-that/compare/v0.5.0...master
+[v0.5.0]: https://github.com/smartive/proc-that/compare/v0.4.0...v0.5.0
 [v0.4.0]: https://github.com/smartive/proc-that/compare/v0.3.2...v0.4.0